# hospital-management-api/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hospital Management API",
    "version": "1.0.0",
    "description": "A RESTful API for managing patients, doctors, appointments, medical records, and GitHub OAuth authentication. Developed as part of the CSE 341 Final Project.",
    "contact": {
      "name": "Hospital Management API Development Team"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://final-projectcse341.onrender.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/patients/": {
      "get": {
        "tags": [
          "Patients"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "age": {
                    "example": "any"
                  },
                  "gender": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "address": {
                    "example": "any"
                  },
                  "emergencyContact": {
                    "example": "any"
                  },
                  "diagnosis": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/doctors/": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Doctors"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "specialization": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "department": {
                    "example": "any"
                  },
                  "licenseNumber": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/appointments/": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Appointments"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "patientId": {
                    "example": "any"
                  },
                  "doctorId": {
                    "example": "any"
                  },
                  "appointmentDate": {
                    "example": "any"
                  },
                  "appointmentTime": {
                    "example": "any"
                  },
                  "reason": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "notes": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/medical-records/": {
      "get": {
        "tags": [
          "Medical Records"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Medical Records"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "patientId": {
                    "example": "any"
                  },
                  "doctorId": {
                    "example": "any"
                  },
                  "diagnosis": {
                    "example": "any"
                  },
                  "symptoms": {
                    "example": "any"
                  },
                  "treatment": {
                    "example": "any"
                  },
                  "medications": {
                    "example": "any"
                  },
                  "notes": {
                    "example": "any"
                  },
                  "recordDate": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Express session cookie created after successful GitHub OAuth authentication."
      }
    },
...
  },
...
}